got rid of print statements in the commands

+++ src/backEnd/commands/For.java
@@ -6,8 +6,8 @@ import java.util.LinkedList;
 public class For extends Command {
 
   private final int number = 2;
-  private int start;
-  private int stop;
+  private double start;
+  private double stop;
 
   public For(){
     super();
@@ -15,13 +15,13 @@ public class For extends Command {
   }
   public For(LinkedList<String> varags, Control control){
     super(varags,control);
-    start = Integer.parseInt(varags.get(0));
-    stop = Integer.parseInt(varags.get(1));
+    start = Double.parseDouble(varags.get(0));
+    stop = Double.parseDouble(varags.get(1));
     System.out.println(start + " " + stop);
   }
 
   @Override
-  public int repeatCom() {
+  public double repeatCom() {
     return (stop-start);
   }
 }
