+++ src/backEnd/commands/Repeat.java
@@ -5,7 +5,7 @@ import java.util.LinkedList;
 
 public class Repeat extends Command {
 
-  private int repetition;
+  private double repetition;
   private Control c;
   private static final int number =  1;
 
@@ -17,11 +17,11 @@ public class Repeat extends Command {
     super(varargs, control);
     super.setControl(control);
     c = control;
-    repetition = Integer.parseInt(varargs.get(0));
+    repetition = Double.parseDouble(varargs.get(0));
   }
 
   @Override
-  public int repeatCom(){
+  public double repeatCom(){
     return repetition;
   }
 }
